refactor(views): split god view files into per-concern files
Verbatim moves, no behavior change. File-scope private promoted to internal
only where references now cross files. Synchronized pbxproj groups pick up the
new subdirectories automatically.

- UsageDashboardView.swift 1618 -> 409 + Views/Dashboard/ (7 files)
- MenuBarView.swift 1025 -> 152 + Views/MenuBar/ (5 files)
- SettingsView.swift 775 -> 543 + Views/Settings/ (2 files)

Verified: swift build clean, swiftlint --strict 0.
