+++ .agents/SESSIONS/2026-07-02.md
@@ -103,8 +103,32 @@ start; the audit was done inline instead.
 ### Next steps
 
 - Watch PR CI: read actual coverage % from the test job log, ratchet
-  `COVERAGE_THRESHOLD` in the same PR.
-- Follow-up tasks (spawned as chips): view splits, MeterBarShared extraction
-  (Mac Studio / full Xcode), release signing + notarization.
+  `COVERAGE_THRESHOLD` in the same PR. *(Done — measured 9.0%, floor set to 8.)*
+- Follow-up tasks: MeterBarShared extraction (Mac Studio / full Xcode), release
+  signing + notarization.
 - Tests for `UsageDataManager` orchestration and `CostTracker` scan paths remain the
   biggest coverage gap.
+
+---
+
+## Session 2: View-file split (R8)
+
+**Status:** Complete — PR stacked on #66, branch `claude/view-split`
+
+Split the three god view files into per-concern files (verbatim moves; file-scope
+`private` promoted to internal only where references now cross files). The Xcode
+project uses file-system-synchronized groups, so the new subdirectories join the
+app target automatically.
+
+- `UsageDashboardView.swift` 1,618 → 409 + `Views/Dashboard/` (7 files:
+  WindowController, DashboardModels, DashboardCards, CostScanViews,
+  DailyUsageChart, CostBreakdownViews, DashboardHelpers)
+- `MenuBarView.swift` 1,025 → 152 + `Views/MenuBar/` (5 files: ProviderLogo,
+  PopoverOverview, ResetCountdown, ExtraUsageStatusPill, UsageBar)
+- `SettingsView.swift` 775 → 543 + `Views/Settings/` (SettingsComponents,
+  AdminKeyHelp) — the remaining 543 lines are the cohesive settings Form itself.
+
+Verified with `swift build` (the root SwiftPM package compiles Views) and
+`swiftlint lint --strict` (0). Split done mechanically with line-range scripts
+(kept in session scratchpad); two splitter artifacts caught by build/lint and
+fixed (un-stripped `private` on shared helpers; double blank lines).
