+++ MeterBar/Views/SettingsView.swift
@@ -541,235 +541,3 @@ struct SettingsView: View {
         }
     }
 }
-
-private struct SettingsPanelSection<Content: View>: View {
-    let title: String
-    let logoKind: ProviderLogoKind?
-    let systemImage: String?
-    let color: Color
-    let content: Content
-
-    init(
-        title: String,
-        logoKind: ProviderLogoKind,
-        color: Color,
-        @ViewBuilder content: () -> Content
-    ) {
-        self.title = title
-        self.logoKind = logoKind
-        self.systemImage = nil
-        self.color = color
-        self.content = content()
-    }
-
-    init(
-        title: String,
-        systemImage: String,
-        color: Color,
-        @ViewBuilder content: () -> Content
-    ) {
-        self.title = title
-        self.logoKind = nil
-        self.systemImage = systemImage
-        self.color = color
-        self.content = content()
-    }
-
-    var body: some View {
-        Section {
-            content
-        } header: {
-            HStack(spacing: 6) {
-                if let logoKind {
-                    ProviderLogoView(kind: logoKind, size: 14, foregroundColor: color)
-                } else if let systemImage {
-                    Image(systemName: systemImage)
-                        .foregroundStyle(color)
-                }
-                Text(title)
-            }
-        }
-    }
-}
-
-private struct SettingsRowView<Content: View>: View {
-    let title: String
-    let detail: String?
-    let content: Content
-
-    init(title: String, detail: String? = nil, @ViewBuilder content: () -> Content) {
-        self.title = title
-        self.detail = detail
-        self.content = content()
-    }
-
-    var body: some View {
-        LabeledContent {
-            content
-        } label: {
-            VStack(alignment: .leading, spacing: 2) {
-                Text(title)
-                if let detail {
-                    Text(detail)
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                }
-            }
-        }
-    }
-}
-
-private struct SettingsNotice: View {
-    let text: String
-    let color: Color
-
-    var body: some View {
-        Text(text)
-            .font(.caption)
-            .foregroundStyle(color)
-            .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-private struct SettingsDivider: View {
-    var body: some View {
-        Divider()
-    }
-}
-
-private struct StatusPill: View {
-    let title: String
-    let isConnected: Bool
-
-    var body: some View {
-        Label(title, systemImage: isConnected ? "checkmark.circle.fill" : "circle")
-            .foregroundStyle(isConnected ? MeterBarTheme.success : Color.secondary)
-            .font(.subheadline)
-    }
-}
-
-private struct AccountProfileRow: View {
-    let account: ClaudeCodeAccount
-    let onRemove: () -> Void
-
-    var body: some View {
-        HStack(spacing: 10) {
-            Image(systemName: account.isDefault ? "person.crop.circle" : "person.crop.circle.badge.plus")
-                .foregroundStyle(MeterBarTheme.claudeAccent)
-                .frame(width: 18)
-
-            VStack(alignment: .leading, spacing: 2) {
-                Text(account.name)
-                    .font(.subheadline)
-                    .fontWeight(.semibold)
-                Text(account.configDirectory ?? "Default Claude CLI profile")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-
-            Spacer()
-
-            if !account.isDefault {
-                Button("Remove", role: .destructive, action: onRemove)
-                    .buttonStyle(.bordered)
-            }
-        }
-        .padding(.vertical, 4)
-    }
-}
-
-/// Shared admin-key help sheet. The Claude and OpenAI variants only differ by
-/// copy and the console URL, so they share one layout.
-private struct AdminKeyHelpView: View {
-    @Environment(\.dismiss)
-    var dismiss
-
-    let title: String
-    let intro: String
-    let steps: [String]
-    let note: String
-    let consoleButtonTitle: String
-    let consoleURL: String
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Text(title)
-                .font(.title2)
-                .bold()
-
-            Text(intro)
-                .foregroundColor(.secondary)
-
-            Divider()
-
-            VStack(alignment: .leading, spacing: 8) {
-                ForEach(Array(steps.enumerated()), id: \.offset) { index, step in
-                    Text("\(index + 1). \(step)")
-                }
-            }
-
-            Divider()
-
-            Text(note)
-                .font(.caption)
-                .foregroundStyle(MeterBarTheme.warning)
-
-            HStack {
-                Button(consoleButtonTitle) {
-                    if let url = URL(string: consoleURL) {
-                        NSWorkspace.shared.open(url)
-                    }
-                }
-                .buttonStyle(.borderedProminent)
-
-                Spacer()
-
-                Button("Close") {
-                    dismiss()
-                }
-            }
-        }
-        .padding()
-        .frame(width: 500, height: 400)
-    }
-}
-
-struct ClaudeHelpView: View {
-    var body: some View {
-        AdminKeyHelpView(
-            title: "How to get Claude Admin API Key",
-            intro: "The Usage API requires an Admin API key, which is different from a regular API key.",
-            steps: [
-                "Go to the Claude Console",
-                "Navigate to Settings → Admin Keys",
-                "Click 'Create Admin Key'",
-                "Copy the key (starts with sk-ant-admin...)",
-                "Paste it in the field above"
-            ],
-            note: "Note: You must be an organization admin to create Admin API keys. "
-                + "Individual accounts cannot access the Usage API.",
-            consoleButtonTitle: "Open Claude Console",
-            consoleURL: "https://console.anthropic.com/settings/admin-keys"
-        )
-    }
-}
-
-struct OpenAIHelpView: View {
-    var body: some View {
-        AdminKeyHelpView(
-            title: "How to get OpenAI Admin API Key",
-            intro: "The Usage API requires an Admin key from your organization settings.",
-            steps: [
-                "Go to OpenAI Platform",
-                "Navigate to Settings → Organization → Admin Keys",
-                "Click 'Create new admin key'",
-                "Copy the key",
-                "Paste it in the field above"
-            ],
-            note: "Note: You must be an organization owner or admin to create Admin keys.",
-            consoleButtonTitle: "Open OpenAI Settings",
-            consoleURL: "https://platform.openai.com/settings/organization/admin-keys"
-        )
-    }
-}
